[IMP] point_of_sale: refactor set table.
In order to use the blackbox, we don't use the synchronisation of the table.
So we have to change the way of how set table is done to isolate the syncing part from the original part of setting table.
This change allow us to be more flexible to override the behavior.