[IMP] point_of_sale: Split bill
Because the blackbox doesn't allow the removing of lines, we have to change the function in this case. The previous implementation doesn't allow us to be flexible and leads us to incorrect behavior while dealing with multiple line of the same article.
Those changes allow us to handle all of this easier without affecting the current behavior in the normal use cases.