[IMP] portal: improved an attachment for portal chatter
before this commit: an attachment doesn't support document preview(quick view)
while clicking on an attachment

in this commit, improved portal chatter for messaging and attachments.
moved/update the code related to the document viewer in common assets for
backend and frontend. also, the appearance of attachment is changed to what it
looks like in the backend.

task-2054662

Co-authored-by: jpr-odoo <jpr@openerp.com>