[IMP] website: introduce cookies bar on website
With this commit, it is now possible to activate a cookies bar on a website
through the res.settings.

The cookies bar can be edited in the frontend by entering edit mode (to be
editable, it will be shown even if the user has hidden the cookies bar).

The cookies bar will appear until the user clicks on 'I agree'.

task-2087003