[IMP] sale_timesheet: Allow to fold/unfold project overview
Allow to fold/unfold the details by SO. It should be folded by default.

Note: Although qweb views have a generic mechanism to manage fold and unfold, this standard
behavior is not used here.
It would imply refactoring the entire code preparing the template data (700+ lines of
code which are hard to read). The quick and dirty way is used for now.
(I wish good luck for the dev if it's one day necessary.)

Task 2119567