[IMP] web, *: Change placeholder by default when record is unassigned
TaskID: 2060206

We want to have a fa-user-slash icon when a record is unassigned instead
of 'placeholder.png'
A method is created in the BaseModel to have a generic method to change
easily the placeholder for other models.

[FIX] kanban test to keep the same behaviour.
Attention the behaviour is a bit different.
Because, now the default image it's give by the server to
change easily the default image when a record doesn't have an image.
Thus, we don't say if it's the default placeholder, but we can say it's
not the same image of the record (in this test, the record, it's the
partner).