[IMP] mail: add author field on mail template
When you configure a mail template, you can specify an author for the mail. This allows to simplify
the setting of a template because you no longer need to now the format to select the email of the author.
Example: for an invoice, you can set the author to "${object.user_id}" and the email_from field would be set to
"${object.user_id.email_formatted |safe}".
Additionaly, this allows when you send the mail to have the good author and not the user who clicks on the button send.

task-2091859