[FIX] hr_expense : fix issues in hr_expense in a multicompany situation.
The company_id wasn't propagating correctly from the expense or expense report
to the invoice and payment.

The company_id wasn't being propagated correctly when creating an expense
from an expense report.

The employee_id was prefilled automatically with one linked to the user but the company wasn't
taken into account.
This caused issue when a manager wanted to create an expense for an employee in a different company.

This commit fixes those issues and also introduces a unit test to ensure that it keeps working.

Task-2088891