[IMP] mail: visible the attachment counter with the attachment box
Currently, the attachment counter is only visible when the 'message_ids' field
is available in the view. Actually the counter should always be visible whenever
the attachment box is visible in the chatter.

Thus, visible the attachment counter with its box for all mailFields instead of
only 'message_ids'.

Also, remove the class 'o_chatter_composer_active' while closing the attachment
box as it is putting the border which is giving the ugly UI.

task-2149348

Closes https://github.com/odoo/odoo/pull/42672