[FIX] account: compute of amount residual in foreign currency
When having an invoice in foreign currency and a partial payment in
domestic currency (in a different date with a different exchange rate),
the computed amount due is incorrect. When computing the
amount_residual_currency the payment (in domestic currency) is converted
to the foreign currency using the exchange rate of the invoice date, and
not the one of the payment date (see 56d92213b82d417e48cb5a3432eb57dcba706b6e).

If the account.partial.reconcile line will have a currency_id and a
amount_currency the amount_residual_currency will not be converted but
simply added.

This behaviour was introduced with e1306066026649acd388b49d7429337d87cf819a
and backported to version 11.0 with 3363d7235137f286458c734a1a0417beeb1f66bb.
In version 12.0 this behaviour was partially reverted in d3d26120614139fd7d7e888bd66d21de5158a034.

opw-2156001