[REF] FSM: Make buttons visible based on fields
Make buttons visible based on computed fields instead of attributes

The various buttons in FSM are currently displayed based on a set of attributes.
As there are many different configurations possible for projects, it makes a lot of conditions
This task aims at replacing these attributes by computed fields in order to gain clarity and to avoid the various bugs we recently faced.

TASK-ID: 2126857