[IMP] purchase_stock: picking return - perf
Before this commit:
When doing a return on a picking with multiple hundreds of products,
the return processing would timeout.

This commit does in one SQL request what was done with multiple .mappeds
The .mapped() made the business logic very clear but were also
prefetching a lot of useless data.

Empirical testing shows a performance gain of between 5 and 7x on large
datasets.

opw-2161341