[IMP] api: move cache_key to the Environment and cache it
Computing the `cache_key` turned out to be a big factor during the
lifespan of a `BaseModel.mapped` call and a lot of this time is spent
computing the same `cache_key` over and over.

These unnecessary computations can be easily reduced to a couple by
using python's `functools.lru_cache` which will keep a cache per
environment that depends only on the field for which the `cache_key` is
being computed.

The `maxsize` of the cache is not chosen arbitrarily:
    - CPython's `functools.lru_cache` is optimized for power-of-two
    maxsizes.
    - It is unlikely that a single business flow will access many more
    different fields than the amount defined by maxsize, however
    this means that if there are more different field accesses than
    `maxsize`, the cache is practically useless iteration-to-iteration.

The result of this patch is up to 50% faster `Field.__get__` which in
turn means a GLOBAL gain in performance, especially for methods /
functions that rely heavily on `__get__` such as `BaseModel.mapped`