[FIX] mrp: truncate millisecond to check overlaps
In some flows, a planned_datetime_start/finished will contains
milliseconds. In these cases, we can detect overlaps/conflicts between
workorders but these conflicts can't be see in the web interface
(because datetime is round when it is display).
Then, change the conflicts computation to remove milliseconds
in search of conflicts.

task-2169447